test(pg): append-session-event split into nine files, the live suite requires the test database instead of skipping (issue 878)
append-session-event.test.ts held 62 unit tests and an 8-test live Postgres
suite in one file. The live suite read OPENBRAIN_TEST_DATABASE_URL itself and
swapped in describe.skip when the variable was absent, so an unconfigured run
reported 62 pass / 9 skip and exited 0 -- indistinguishable at the exit code
from a run that exercised the Postgres behavior.

The file is retired and its tests are redistributed across nine files, each a
coherent subject, each under the file size the lint config allows:

  append-session-event-auth.test.ts            auth and lane lookup
  append-session-event-lane-creation.test.ts   create_if_missing paths
  append-session-event-lane-state.test.ts      lane status handling
  append-session-event-failure-paths.test.ts   embed and persist failures
  append-session-event-provenance.test.ts      created_by and provenance
  append-session-event-share-gate.test.ts      sync share-nomination gate
  append-session-event-share-lifecycle.test.ts share lifecycle actions
  append-session-event-citation.test.ts        database errors and citations
  append-session-event.pg.test.ts              the live Postgres suite

append-session-event-test-helpers.ts carries the mock pools and the tool-client
setup the halves share. The live suite now calls requireTestDatabaseUrl() at
module scope: absent the variable it throws test_database_required and the run
fails loudly rather than reporting a green skip.

scripts/assert-db-tests-ran.ts recorded minTests 4 for that suite while it
emits 8. The entry is corrected to the count JUnit reports, and
MIN_TOTAL_LIVE_TESTCASES moves 328 -> 332 by the same delta with a clause added
to its existing chain, so the global floor cannot fall behind the per-suite one.

No test name or assertion changed. Each it body was hoisted mechanically to a
module-scope function so max-lines-per-function is measured per test rather
than at the describe callback; the `x!` non-null assertions became
expectDefined(), and the result-shape casts became the shared getErrorText()
and parseToolResult() helpers.

The eight steps of the split were verified on the working tree in sequence
during session 12, and the whole tree was re-checked before this commit:
oxlint --deny-warnings clean on all eleven files, bunx tsc --noEmit exit 0,
bun run test:isolated exit 0 at 4010 pass / 0 fail across 289 files, and the
nine append-session-event files together at 70 pass / 0 fail / 0 skip -- the
62 unit tests and 8 live tests the original carried, none lost.

Refs issue 878.
